backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
matplotlib
seaborn
openai
langgraph
python-dotenv
reportlab
